backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
influxdb-client==1.37.0
redis==5.0.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
twilio==8.10.0
typing-extensions==4.8.0
python-socketio==5.10.0
httpx==0.25.2